Modularize Code (#26)
* initial stab modularizing code
main app class deriving from Sinatra::Base
seperate route/controller classes for icons, identity and api namespaces
move helpers to own class
